sanskrit-ocr-workspace/requirements.txt
numpy
pandas
opencv-python
scikit-learn
tensorflow
torch
torchvision
Pillow
matplotlib
tqdm
PyPDF2
imgaug
seaborn
jupyter
notebook